+++ src/main/kotlin/codegen/FlutterXPluginInfo.kt
@@ -5,7 +5,7 @@ object FlutterXPluginInfo {
     const val CHANGELOG: String = """
 <h2>5.6.2 - 2025-05-08</h2>
 
-<ul><li>新增 dio重启逻辑</li><li>优化 l10n编辑面板，新增 preview text 选项，在 tree中预览文本</li></ul>
+<ul><li>Add <code>dio</code> restart button (settings)</li><li>Optimize <code>l10n</code> editing panel, add preview text option, preview text in tree (l10n panel)</li><li>Add &quot;delete l10n key&quot; action, in right popup menu (l10n panel)</li><li>Add &quot;Rename l10n key&quot; action, in right popup menu (l10n panel)</li><li><a href="https://github.com/mdddj/dd_flutter_idea_plugin/blob/master/CHANGELOG.md">full changelog</a></li></ul>
 """
 
 }